requirements.txt
pillow==7.2.0
requests==2.24.0
